requirements.txt
kiteconnect>=5.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
rich>=13.7.0
asyncio>=3.4.3
